backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==2.5.1
pymongo==3.12.3
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pydantic-settings==2.1.0
pydantic[email]==2.11.1
python-dotenv==1.0.0
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
aiofiles==24.1.0
python-dateutil==2.8.2
